Time: 0 ms (100.00%) | Memory: 8.6 MB (20.88%) - LeetSync

+++ 7-reverse-integer/reverse-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int reverse(int x) {
-        long ans = 0;
+        int ans = 0;
         while(x!=0){
             int rem = x %10;
             if(ans > INT_MAX/10 || ans < INT_MIN/10) return 0;
